Fixed DokanUnmount bug
As reported in the mailing list, DokanUnmount called
DokanRemoveMountPoint with an extra character at the end. Fixed it, and
mention DokanRemoveMountPoint in the readme.txt, since is an exported
function in the public header and is necessary if you mount the file
system in a path, not a drive letter.

See:
  https://groups.google.com/d/topic/dokan/zAZukiICdNk/discussion