Use _malloca() for debug strings to avoid buffer overflows
While the sprintf_s functions are supposed to protect from buffer overflows and
other errors, they still raise an exception if the buffer is too small, so any
debug string longer than 512 characters still crashes the program. In practice,
this limit is exceeded very easily, for example, when matching a file name
with a length of MAX_PATH against itself.

Instead of just increasing the buffer size and/or using snprintf, I prefer to
use a combination of _vscwprintf() and _malloca()/_freea(), which computes the
final length and then calls either alloca() for strings below 1024 bytes or
malloc() for longer ones. This makes sure the string always gets shown in full
(well, except for the unlikely case of a malloc() failure) and should eliminate
this bug class without impacting performance too much.