The CreateFile callback now resembles ZeCreateFile() instead of CreateFile()
The previous implementation of CreateFile() was not passing a lot of the
information that gets passed into the driver. In order to provide driver
writers with as much information as possible I modified the CreateFile
to more closely resemble the parameters of ZwCreateFile().

I added some interop structs for moving some kernel data to user mode.
Added missing kernel mode FILE_* flags
Rearranged headers to make fileinfo.h and public.h available to user
mode drivers