Add CreateFileEx to DOKAN_OPERATIONS
This commit adds a new operation CreateFileEx() to DOKAN_OPERATIONS. The operation can be used to create/open files/directories. The commit does not remove CreateFile/CreateDirectory/OpenDirectory as to not break existing file systems.

The commit introduces the following operation:

    NTSTATUS (DOKAN_CALLBACK *CreateFileEx) (
        LPCWSTR,                // FileName
        DWORD,                  // DesiredAccess
        DWORD,                  // ShareAccess
        DWORD,                  // CreateDisposition
        DWORD,                  // CreateOptions
        DWORD,                  // FileAttributes
        PSECURITY_DESCRIPTOR,   // Reserved
        PULONG,                 // &IoStatusBlock.Information
        PDOKAN_FILE_INFO);

This operation is different from the existing CreateFile as follows:
- CreateDisposition uses the constants from NtCreateFile:
-- FILE_SUPERSEDE, FILE_CREATE, FILE_OPEN, FILE_OPEN_IF, FILE_OVERWRITE, FILE_OVERWRITE_IF
- CreateOptions uses the constants from NtCreateFile:
-- FILE_DIRECTORY_FILE, FILE_NON_DIRECTORY_FILE, etc.
- FileAttributes contains only a (new) file's attributes. It is not FlagsAndAttributes as in Win32 CreateFile.
- It has a Reserved parameter that can be used in the future to pass a security descriptor to attach to a (new) file.
- It accepts a pointer to IoStatusBlock.Information where the file system must place information about the result of the operation (refer to NtCreateFile documentation):
-- FILE_CREATED, FILE_OPENED, FILE_OVERWRITTEN, FILE_SUPERSEDED, FILE_EXISTS, FILE_DOES_NOT_EXIST

The main advantage of using CreateFileEx is that is passes more information to the file system so that it can make correct decisions with respect to flags such as FILE_DIRECTORY_FILE, FILE_NON_DIRECTORY_FILE. It also eliminates the split between CreateFile/CreateDirectory/OpenDirectory.