Replaced the NSIS installer with WiX
Added a new powershell based build script that properly reads the
registry for Visual Studio's location
NOTE: To properly install a certified WHQL driver a CatalogFile needs to
be provided and Legacy="yes" needs to be set to "no" in the x64/x86
Product.wxs files
Version numbers need to be updated in each WiX installer, the
bootstrapper, Dokan.BootstrapperUX, and finally dokan.inf. Find in Files
is your best friend here.
Still a lot of testing required on various OS's and x86