Fix `CreateDisposition` checks in the Dokan Fuse wrapper
CreateDisposition is not remapped from kernel constants in `wdm.h`
(`FILE_CREATE`, `FILE_OVERWRITE`...) to user values in `fileapi.h`
(`CREATE_NEW`, `CREATE_ALWAYS`...), but the checks are performed with the user
constants. The values are not the same, and incorrect checks break the file
creation logic.

The fix uses translation rules defined in the
`DokanMapKernelToUserCreateFileFlags` method from `dokan.c`.