Set common last-error codes in the `ZwCreateFile` callback
According to Dokan documentation, in case a call to `CreateFile` with a
`CreateDisposition` set to `FILE_OPEN_IF` or `FILE_OVERWRITE_IF` is opening an
already existing file successfully, we have to set the last error to
`ERROR_ALREADY_EXISTS`. According to Windows development documentation for the
`CreateFile` function, for a similar case with a disposition set to
`FILE_CREATE`, we have to set the last error to `ERROR_FILE_EXISTS`.

According to Windows development documentation for the `CreateFile` function, in
case a call to `CreateFile` with a `CreateDisposition` set to `FILE_OPEN` or
`FILE_OVERWRITE` can’t find a file to open or truncate, we have to set the last
error to `ERROR_FILE_NOT_FOUND `.