Allow to delete files and directories on FUSE file systems
Current DokanFUSE implementation doesn’t always call the `unlink` or `rmdir`
functions. It also doesn’t follow Dokan’s requirements on how to handle
`Delete*` callbacks.

The fix is to check whether it’s possible to delete a file in the `DeleteFile`
or `DeleteDirectory` callbacks, return appropriate status codes, and call the
FUSE `unlink` or `rmdir` functions in a consequent `Cleanup` call when the flag
`dokan_file_info->DeleteOnClose` is set to `true`.