dokan_fuse: Use %ls for wide-character debug-print
In dokan_fuse debug output the filenames are now printed out using %ls
in order to be compliant with the C++ standard and thus make the
printing work when compiling with GCC/MingW.

MSVC will happily print a wide-character-string when using %s but will
fail to print out single-byte strings. This has to be kept in mind
when printing out single-byte strings using wide-character fwprintf.

MSVC14 supports defining _CRT_STDIO_ISO_WIDE_SPECIFIERS to bring their
format-string handling in line with the standards.