dokan_wix: Fix KB3033929-check
Due to some oddity, we need to specify a version that is one version
lower than the one we are actually searching for.
References: #260, #244, #187

Quoting WiX-documentation:
Important: When doing a locale-neutral search for a file, you must set
the MinVersion property to one revision number lower than the actual
version you want to search for.