Appveyor: Upgrade all Cygwin packages before build
Upgrading all packages ensures that we are running the latest Cygwin.
Sometimes dependencies of packages are incorrect and dependant packages
are thus not upgraded automatically when they should be.

This makes the Appveyor build working again after the cmake-package in
Cygwin was upgraded a few days ago breaking our build. Closes #335.