dokan_fuse:Use macros for printing DWORD and ULONG
DWORD and ULONG are defined as unsigned long int on Windows. Cygwin64 is
LP64. Thus it defines them as unsigned int (to keep them 32-bits wide).
In order to enable portable printing, we define macros modeled after
inttypes.h like for example PRIuDWORD or PRIxDWORD.

Using these macros ensures that the correct printf-format-specifier is
used depending on the platform.