CI: Do not run Coverity on forks and pull requests
If the secret environment variable with the Coverity credentials is not
available, we output a message and return instead of trying to run
Coverity. The previous check only checked for pull-requests, instead we
now check whether the credential is available or not, which checks for
the actual cause and thus catches forks as well.