Issue #490:
Do not add trailing backslash in front of alternate data
stream name (which begins with colon).

If the alternate data stream is a stream of the root directory, then
do not prepend the relatedFileName (which is a backslash) to the file
name (which is the stream name, which begins with a colon).  This is
to enable alternate data steams of the root directory itself to be
opened.