FUSE: cross-compile 32-bit Cygwin DLL from 64-bit
Cygwin64 can compile binaries for Cygwin32. Install the corresponding
cross-compiler and use it for compiling dokan_fuse.

This should save quite a bit of time spent downloading 32-bit Cygwin
updates on AppVeyor. It also makes live easier, because there is no
longer any need to install the 32-bit Cygwin in parallel to the 64-bit
one on local machines.