MemFS - Add a new FS sample project: dokan_memfs
MemFS is a better example to debug and know the dokan driver/library feature supported and NTFS compliant.
The FS pass most of WinFSTest and IFSTest. It looks to be stable enough to be included in the driver.
It hasn't been test with real usage but it is expected to run without issue.
MemFS is written in c++ and is under MIT license.