CI - Remove Coverity
Coverity no longer accept to analyze the build as we are under the 85% capture success.
The % dropped when adding spdlog submodule for memfs. Coverity has some options to exclude directories from the capture count but not for headers...as spdlog is header only, there is no options to exclude the lib.
Coverity is therefore unusable for now.