Kernel - Deep debug log improvement.
The current logging is unusable to understand and track concurrent IO paths. The new logging macros focus on resolving this issue by printing the current IRP and it's attached FileObject until completion.

There are two main log functions DOKAN_LOG_BEGIN_MJ & DOKAN_LOG_END_MJ that take care of displaying all the Irps information on begin and completion dispatch. DOKAN_LOG_IRP will take care of logging attached to the Irp in the meantime. DOKAN_LOG is now the default logging function to use outside an Irp context. Note: There are still DOKAN_LOG that can be converted to DOKAN_LOG_IRP but will be addressed in another CL.

A couple of new functions are added in log.c to help convert the kernel values to string. This allows us to remove a lot of print logs like DDbgPrint("    FSCTL_REQUEST_FILTER_OPLOCK\n"); all over the code base.

The filename is now only printed during CreateFile with the FileObject. Later only the FileObject is printed for avoiding flooding the log with possible long paths.

All the added functions and defines are nullified on release.