Dokan - Allow kernel driver logs to be dispatched to userland.
DOKAN_EVENT_DISPATCH_DRIVER_LOGS is a new option at mount time that make the driver start to cache global & specific volume logs in a global cache. When the cache reach the DOKAN_LOG_POP_ENTRY_START limit and an IRP to a volume is logged, a EVENT_CONTEXT of EVENT_CONTEXT_MAX_SIZE with the new custom Major IRP DOKAN_IRP_LOG_MESSAGE is created, filled with logs and dispatched to userland.