Kernel - Add an option to use FSCTL Event type instead of IOCTL with dwDesiredAccess nullified.
The actual dwDesiredAccess used to communicate with Dokan is not needed if FSCTL events are used.

The CL duplicates all existing IOCTL events to their FSCTL version. While using IOCTL, the DeviceObject for the request is type of VCB while when FSCTL is used, the DeviceObject is a DCB type. This CL has a small hack that creates a temporary REQUEST_CONTEXT that is built like it is coming from a VCB request during one of the new FSCTL DCB requests. This should be changed when we fully move to FSCTL.

Note: FSCTL_NOTIFY_PATH & FSCTL_ACTIVATE_KEEPALIVE are existing FSCTL that actually target the VCB with dummy files. It might be good to also move them to DCB and avoid having unusable files.

https://github.com/dokan-dev/dokany/issues/953