kernel - Fix a very rare race condition that make library fail to detect unmount
What is happening is that the device is unmounted during the moment where a coming EVENT_WAIT IRP is appended to the PendingEvent list that was already emptied by MoveIrpList by the unmount process. The IRP would have just been lost in this list and never canceled.

Adding a check when we have the list locked to be sure there is no unmount happening.