Kernel - Fix invalid buffer size count when PREPARE_OUTPUT is used with types ending with a dynamic size field that are later filled with AppendVarSizeOutputString.
PREPARE_OUTPUT will use the sizeof(struct) to set the output used space. If the struct is ending with a dynamic size field like FILE_NAME_INFORMATION.FileName[1] the FileName[1] we should not count it and let AppendVarSizeOutputString later add it if it gets set. It is valid to return a buffer of size FIELD_OFFSET(FILE_NAME_INFORMATION, FileName) without anything set in FileName in purpose to get the FileNameLength to know the correct buffer to provide for the next try.

This commit is fixing WSL version 1 mount that is failing with value too large for defined data type WSL has a strict workflow where it opens and query FileNameInformation for the root \ with an output buffer of size of 6 == FIELD_OFFSET(FILE_NAME_INFORMATION, FileName) + WCHAR for \ and expect us to exactly fill it and set the used space to 6 and not 8 == sizeof(FILE_NAME_INFORMATION)