Dokan - Remove unsupported `FileNetworkPhysicalNameInformation` QueryInformation
The current code does not concat the network path (Should be using FillNameInformation).
The FileNameLength is not set. We do fill the buffer with the FileName but FileNameLength will always be 0.

This makes me think we really do not need to support this query, otherwise we would have received reports for miss usage.