Kernel - Fix relative rename with Volume device as RootDirectory BSOD
Renaming a file when the volume if shared with SMB net use x: \\localhost\C$\Users\liryna\MountPoint (C:\Users\liryna\MountPoin as mount point folder) is proceed as relative rename with the RootDirectory being the volume device (not the volume root \). In that case, there is no DokanCCB set to FsContext2 for the RootDirectory handle as we do not set one during IRP_MJ_CREATE. As the full destination path is provided by TargetFileObject, we just fallback to the fully qualified rename.