Kernel - Reset top-level IRP when creating / deleting reparse points.
Setting top-level IRP is a signal to the Filter Manager that locks may be held above it. This makes filter drivers (like ProjFS) to not be able to query names when we call them during the creating / deleting reparse point.
Therefore in that case ProjFS failed our request.

Since we are not holding any lock and have set the top-level IRP ourselves, we are safe to temporarily reset the thread field.