Revert "CI: Install MSYS2 on VS2019 image"
This reverts commit f980bb2bfb504dcea7538af2587c23492ce63ef2.
MSYS2 is by now installed in the "Visual Studio 2019" image on AppVeyor.
Thus, we can simply update the existing installation like it was done
previously.

As a side effect, this also avoids exceeding AppVeyor's maximum cache
size, which recently became a problem because substantially less data
needs to be cached.