Library - Correctly set PullEventTimeoutMs value and Make MainPullThread to not process events
In some cases, the PullEventTimeoutMs would not be set on pool thread (not MainPullThread) which would make them wait indefinitely in the Kernel. That's not a deadlock, they would be awakened later than we expect and during that time they would not be available to process queued events.

The MainPullThread will only pull & dispatch events to the pool to avoid having it deadlock when completing a request #1058