Kernel - Remove eject media support (deadlock)
DiskDeviceControl is called with the RemovalLock acquired during a IOCTL_STORAGE_EJECT_MEDIA (Explorer eject device). DokanUnmount function will also acquire a lock on RemovalLock which will end up with a deadlock as we wait for all locks to be released to finalize the unmount.
Instead of trying to fix this since most fs users might want to keep control of how the FS is ejected, it is better to just remove the code. If someone wants to allow ejecting from a context menu, it can still create a custom one that calls the Dokan API for unmount.