Kernel - Correct locking order for RemoveMountEntry
Fixes #1191

RemoveMountEntry lock order was incorrect. It is not safe to get the global lock after the entry was fetched because a concurrent FindMountEntry holding the global lock cloud be waiting on the entry to be locked while we would do it the other way which would lead to a deadlock.