Kernel - Add `Fcb::UncleanCount` to track and unsure `IRP_MJ_CLEANUP` is executed
In a crash we see that the notify list has a reference to an `Fcb->Filename` which is now free. This shouldn't happen if the `Ccb` was cleaned up correctly. To ensure this is done, this CL added `Fcb::UncleanCount` which is like `Fcb::OpenCount` but only starts to count when Create succeeds and decrements when an attach `Ccb` is cleanup. [Fastfat has it too](https://github.com/microsoft/Windows-driver-samples/blob/89cb2a985612eb83e20b15f6445228ee7f681fba/filesys/fastfat/fatstruc.h#L858-L866). This makes us able to add `assert` in delete Fcb to be sure all our `Ccb` no longer have a reference in the notify list.

That assertion was triggered by a few tests that were doing:
- Cleanup IRP timeout.
- Cleanup IRP canceled.
- Cleanup IRP Oplock completed with a failure.
In those cases we have to force the execution of the cleanup so that the next `IRP_MJ_CLOSE` has the right decremented `UncleanCount` (and also `Ccb` resource released). `IRP_MJ_CLEANUP` is a must succeed (unless pending) IRP as it is not retried when it fails from what I could see. That's why cancelling the IRP does not make it being retried before close is called.