Kernel - Change DeleteOnClose to DeletePending
See https://github.com/dokan-dev/dokany/issues/883
`DELETE_ON_CLOSE` is a handle information that the object should be removed when the last handle is closed. When that handle closes, we transfer it to the Fcb with `DOKAN_FCB_STATE_DELETE_PENDING` which really marks the object as `DeletePending` which means the deletion happens when the last handle is being cleanup.
Note `DOKAN_FCB_STATE_DELETE_PENDING` is directly set during `FileDispositionInformation`.
For userland implementation, this should be transparent beside the variable name change. It is more accurate as they will stop getting `DeleteOnClose` while possibly still having a handle open that prevents them from deleting it.

This is fixing a few ifstests like DeleteOnLastCloseTest

Fixes #998 #883